fix: Prevent metrics from being sent too frequently to OTEL collector
Fixes #250

When using the GKE decoupled deployment, it is possible that a race condition can occur where the Scaler with a 10s periodic export may report metrics more frequently than the OTEL collector's batching interval (also 10s), leading to multiple sets of metrics being pushed to the downstream metrics manager. This can cause Duplicate TimeSeries errors.

Update the counters to push metrics less frequently in periodic export mode, and ensure that metrics are never flushed manually in this mode.

In addition the OTEL collector in unified mode had a 5s default time for batching which is at the limit of the Google Cloud Monitoring requirements that pushes are at least 5s apart.