fix: use a custom service account for Cloud Build
New projects by default do not give the default compute service account
the required roles to run a cloud build. Therefore the cloud functions
do not deploy.

Update the TF scripts to create a build-sa service account and use that
when deploying the Cloud Functions.

Also update versions of Terraform modules to the latest